Proposal for a go-swagger workflow with codegen
Have experimented with the go-swagger tool and found it to be somewhat better than the JavaScript / Node codegen, especially around models and JSON serialisation/deserialisation. This branch has a script code_gen that downlaods the latest YAML from the docs repo, does a bunch of codegen. I'm going to try and see whether this will integrate well with what's already there