Adding gRPC codegen code with example server and client
This is a first step towards consolidating a week or so's thought into the API development for the platform. We're starting from a protobuf definition as our new canonical source of truth(TM), and generating gRPC messages and client/server stubs. These take care of the marshalling and hopefully vaildation in future.

The new scripts/gen_api.sh runs protoc a few times to produce these stubs, as well as a reverse HTTP proxy that maps RESTUL endpoints to gRPC services, thus allowing us to expose REST without hopefully much effort.

The directory structure I've opted for is:

services/
└── trading
    ├── rest-bindings.yml
    ├── trading.pb.go
    ├── trading.pb.gw.go
    ├── trading.proto
    └── trading.swagger.json

The rest-bindings.yml file contains mappings that correlate to the protobuf endpoints, and allow us to add more metadata that is HTTP / REST / Swagger specific.

Apart from the proto file, the rest is generated and shouldn't be edited. The plan would be to next tie this in with the existing service layer code and possibly remove the requirement for the api/ directory.