Decouple services from data
OrderService, TradeService, MarketService:
decouple some of the methods services and expose the logic that
unexported tradingDataService relies on
so it can be used by the scenario runner.
It was the easiest/least disruptive thing to do for now.
Will need to be refactored, unless we go down
the noop blockchain way
in which case those change will become obsolete anyway.