Add vegatime readme
During the preparation for @peterbarrow's Pull Request #1822 we realised that there was no concrete documentation on when to use Vega time and when to use system time. This readme aims to address that.

- Add VegaTime readme

If you feel like this should be in docs rather than a README, let me know. I was conflicted, so I continued the semi-pattern of READMEs in packages, but am aware that's not as discoverable as a docs folder.