Avoid code duplication
Extract a single function dealing with getting orders, fees and
submitting an order so that price monitoring can be applied in
a single place. Note it doesn't fully replicate the old behaviour
as there was no error check after submitting and order in
orderCancelReplace, but that's probably for the best.