Cleaner and uniform market declaration in feature tests
We were using indexes to iterate through the table, which required to update all the indexes when adding or removing a column.

To ease the maintenance of suche table, we introduce a generic layer to get the column by name, with appropriate type conversion.

This makes us switch from the indexes to column name.

To makes things even clearer, we add a specific layer on top of the generic one and abstract away column names into accessors.

Doing so requires us to uniformize the column name across all tests, which is a good thing.