Do not log.Error on ratelimit failure
Ratelimit errors are very frequent in our setup.
In 24hour log file +99% of the lines are related to Ratelimit.

 # Total lines
 ➜ wc j.log
   2351148  40039534 494623216 j.log

 # Lines containing "rate"
 ➜ grep -i "rate" j.log|wc
   2343233 39834826 490907381

 # Lines excluding "rate"
 ➜ logs grep -iv "rate" j.log|wc
   7915  204708 3715835

Also, we can see that for each rate limit error, two modules are logging
the error:

 ➜  grep -i "rate" j.log|head -n2
 ERROR        processor        processor/abci.go:301  Rate limit exceeded          {"key": "2ozoZr2vCx2ZQSfcfpic7WCjAyJh7naiM6eZISeoz4k="}
 ERROR        api.grpc         api/trading.go:153     unable to submit transaction {"error": "public key excedeed the rate limit"}

This PR turns those logs into debug, but also keeps log.Error-ing when a
different type of error happens.