Do not unregister order in applyFees
applyFees is called in SubmitOrder, orderCancelReplace and LeaveAuction.
While it would be acceptable to unregister the order when in the 2 first ones,
it is not when LeavingAuction as this is applied on orders already uncrossed
for which the position has been updated previously in handleConfirmation.

This the unregister order out of the applyFees function, and call it from the
caller.