Panic if a position is not found in the position engine.
Not finding a position in this engine should never be possible,
this would mean that we did not register a position properly, or
that we are trying to unregister one which has maybe already been
unregistered. In any case that's a bug, and shall never happen, a
panic is appropriate in this case.