Improve cache invalidation
We recently improve a lot performances for the indicative price and volume.
Although looking at the pprof profile we can see that lots of time is spent
recomputing the final value, which ultimately would never change unless the
best bid and price range changes, or if anything in the range change.

This change a little bit the cache to:
- behave the same way when in Continuous trading.
- invalidate the cache in Auction, only if the range of bestBid and
  bestAsk was change by any orders or if any order changed in the range