fix: Do not use current time to compute auction duration in update market proposal
This part of the code is already garded by a contidion `currentTime >= enactTime`. So when we are assigning the current time to the close time, it's impossible to get a positive duration whe computing `enactTime.Sub(closeTime)` as the enactment time is equal to current time at best, or behind at worst.

Also, auction duration is ignored anyway, so, we can just keep it this way for now.