fix: Exclude tainted keys from `session.list_keys` endpoint
Excluding tainted keys for this endpoint is done, firstly, by ensuring the permissions do not allow listing those tainted keys.

Listing the keys should be called without wondering wether the list is correct or not. "Correct" means without tainted keys, and with at least one key.

As a result, the wallet is now taking care of enforcing consistency between the permissions and the actual state of the keys inside the wallet.