feat: Move the authentication of wallet API v2 to the transport layer
Before the authentication was part of the JSON-RPC API of the wallet, which was essentially leaking the account management system inside the core business of the wallet. This is not fit for an API that can be integrated by applications that do not have such concept, like browser extensions.

Now the authentication against the wallet API v2 is done through the `Authorization` header, with a custom made `VWT` scheme.