fix: Change the test setup for token store file system updates
Since the events coming from the file system arae not deterministic, especially on different platforms, we had to setup the test in a very peculiar way.

First, it hangs at each store actions, so the file system doesn't squash and flush the changes to the file in one single event. Instead of receiving a single write that only account for the last file update, we are able to retrieve each individual writes.

Then, we verify that, among the FS events we get, we have, at a given step of the test, received the data in a proper state, at least once. If we do, then we move to the next step.