fix: Use vega home to locate snapshot databases
Since the introduction of the Vega home, the file are looked up at a very specific place. This removes the tedious and error-prone path management from the configuration files.

Also, the file management is now pushed down in the technical layer of the LevelDB adapter instead of let it leak in the engine.