fix: Rework snapshot engine for better testability
-> Issue 1

Previously, any call to a newly instantiated snapshot engine that hasn't been cleared and initialized through a call to the method `ClearAndInitialized()` would crash. By design, a newly instantiated engine should be ready to use straight away.

Furthermore, this method was only called externally by tests, which would circumvent the code path running in production, which make the code fragile, and lower our confidence in the unit tests. Hence, the method has been made private, and the tests are now closer to the prodution path.

-> Issue 2

Most of the critical process approving and rejecting a snashot or a snapshot chunk was leaking ouside of the snapshot engine, implemented in the ABCI layer that is notoriously known to be hard to test. Also, this forced the snapshot engine to expose methods so the ABCI could query its state to figure out what to do. This goes against having a self-contained engine that has full control on the state and the decision to take.

-> Issue 3

Restoring state from a local snapshot was not covered by the unit tests, despite being critical.

-> Issue 4

The restoration of the state was hook on the ABCI call `Info()` done by Tendermint, to finalize the hand shake with the rest of the network peer. Based on whether we had some snapshot loaded from state-sync or living locally, the state would be restored, giving priority to state-sync. This had the advantage to have a single point in time where the state would be restored (during the call to `Info()`), but added unecessary complexity, made the call to `Info()` changing state (whereas it's suppose to be a query endpoint), made the core working in the exact opposite way of Tendermint that favor local data over state-sync, and also had the side effect of holding up the node when going throught the steps to join the network because the state was being restored.

-> Issue 5

Instead of resetting the state-sync process, and accept the new snapshot Tendermint offers when Tendermint can't find a way to get all the necesary snapshot chunks, we were returning an "ABORT". Even if Tendermint supports that scenario, there is no reason to not leverage the new offering process.

-> Issue 6

The engine was  not protected against restoring state multiple times in a row, which could cause mayheim.

=> Solution

* Now, the engine is self-contained. All the decisions are taken by the engine and not in the ABCI.
* The tests rely on the same public interface as the production code.
* All the green paths are tested within unit tests.
* The snapshot engine is started right after we start the APIs. If there are some local snapshots, the state is restored at that moment, whether or not state-sync is activated (just like Tendermint does). If state-sync is activated, and no local snapshots is found, the lifecycle of the node is handed over tendermint.
* We accept new snapshots if Tendermint suggests us a new one (because this is usually for good reason).
* The engine can only restore state once per run.