feat: Disallow snapshot state-sync if local snapshots exist
The way Tendermint handles the statesync option is:

- if its set to true and there is no block data locally, performs statesync
- if its set to true and there IS block data locally, it skips the statesync

So, from core point of view, if it had a start-height set and state-sync is on three things can happen:

- there isn't block data, there are also no local vega snapshots, tendermint will start state-sync and it says OK
- there is block data, and there are local vega snapshots, tendermint will NOT statesync and core will load from the snapshot matching block height specified in the configuration
- there ISN't block data, but there ARE vega snapshots, tendermint will try to start state-sync.

In the last situation, it would throw away all existing snapshots and let tendermint enter state-sync. It should instead fail when Tendermint tries to state-sync. It means the core node has got into an inconsistent state where they have unsafe-reset-all'd to get rid of the Tendermint data but we still have core snapshots.

Now,

- it removes all the snapshots and their metadata in the vega command `unsafe_reset_all`.
- if local snapshots are present on the system, the snapshot engine ignores any state-sync, and load the state from the local snapshots.
- the snapshot engine aborts `OfferSnapshot` and `ApplySnapshotChunk` if it started with the local snapshots.